Refactor core and component types in `wasm-encoder`.
This commit simply shuffles types around into a `core` and `component`
submodule and does not affect the exported types in any way.

This mirrors what other crates are doing to delineate core and component
related types.
