+++ crates/wasm-encoder/src/component/aliases.rs
