+++ crates/wasm-encoder/src/component/components.rs
