+++ crates/wasm-encoder/src/component/instances.rs
