+++ crates/wasm-encoder/src/component/modules.rs
