+++ crates/wasm-encoder/src/core/data.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::{encoders, Instruction, Section, SectionId};
 
 /// An encoder for the data section.
 ///
