+++ crates/wasm-encoder/src/core/elements.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::{encoders, Instruction, Section, SectionId, ValType};
 
 /// An encoder for the element section.
 ///
