+++ crates/wasm-encoder/src/lib.rs
@@ -70,359 +70,16 @@
 
 #![deny(missing_docs, missing_debug_implementations)]
 
-mod aliases;
-mod code;
-mod components;
+mod component;
+mod core;
 mod custom;
-mod data;
-mod elements;
-mod exports;
-mod functions;
-mod globals;
-mod imports;
-mod instances;
-mod linking;
-mod memories;
-mod modules;
-mod names;
-mod start;
-mod tables;
-mod tags;
-mod types;
+mod raw;
 
-pub use aliases::*;
-pub use code::*;
-pub use components::*;
-pub use custom::*;
-pub use data::*;
-pub use elements::*;
-pub use exports::*;
-pub use functions::*;
-pub use globals::*;
-pub use imports::*;
-pub use instances::*;
-pub use linking::*;
-pub use memories::*;
-pub use modules::*;
-pub use names::*;
-pub use start::*;
-pub use tables::*;
-pub use tags::*;
-pub use types::*;
-
+pub use self::component::*;
+pub use self::core::*;
+pub use self::custom::*;
+pub use self::raw::*;
 pub mod encoders;
-
-use std::convert::TryFrom;
-
-/// A WebAssembly module section.
-///
-/// Various builders defined in this crate already implement this trait, but you
-/// can also implement it yourself for your own custom section builders, or use
-/// `RawSection` to use a bunch of raw bytes as a section.
-pub trait Section {
-    /// Gets the section's identifier.
-    fn id(&self) -> u8;
-
-    /// Write this section's header and data into the given sink.
-    fn encode<S>(&self, sink: &mut S)
-    where
-        S: Extend<u8>;
-}
-
-/// A WebAssembly component section.
-///
-/// Various builders defined in this crate already implement this trait, but you
-/// can also implement it yourself for your own custom section builders, or use
-/// `RawSection` to use a bunch of raw bytes as a section.
-pub trait ComponentSection {
-    /// Gets the section's identifier.
-    fn id(&self) -> u8;
-
-    /// Write this section's header and data into the given sink.
-    fn encode<S>(&self, sink: &mut S)
-    where
-        S: Extend<u8>;
-}
-
-/// A section made up of uninterpreted, raw bytes.
-///
-/// Allows you to splat any data into module or component.
-#[derive(Clone, Copy, Debug)]
-pub struct RawSection<'a> {
-    /// The id for this section.
-    pub id: u8,
-    /// The raw data for this section.
-    pub data: &'a [u8],
-}
-
-impl Section for RawSection<'_> {
-    fn id(&self) -> u8 {
-        self.id
-    }
-
-    fn encode<S>(&self, sink: &mut S)
-    where
-        S: Extend<u8>,
-    {
-        sink.extend(
-            encoders::u32(u32::try_from(self.data.len()).unwrap()).chain(self.data.iter().copied()),
-        );
-    }
-}
-
-impl ComponentSection for RawSection<'_> {
-    fn id(&self) -> u8 {
-        self.id
-    }
-
-    fn encode<S>(&self, sink: &mut S)
-    where
-        S: Extend<u8>,
-    {
-        sink.extend(
-            encoders::u32(u32::try_from(self.data.len()).unwrap()).chain(self.data.iter().copied()),
-        );
-    }
-}
-
-/// A Wasm module that is being encoded.
-#[derive(Clone, Debug)]
-pub struct Module {
-    bytes: Vec<u8>,
-}
-
-impl Module {
-    /// Begin writing a new `Module`.
-    #[rustfmt::skip]
-    pub fn new() -> Self {
-        Module {
-            bytes: vec![
-                // Magic
-                0x00, 0x61, 0x73, 0x6D,
-                // Version
-                0x01, 0x00, 0x00, 0x00,
-            ],
-        }
-    }
-
-    /// Write a section into this module.
-    ///
-    /// It is your responsibility to define the sections in the [proper
-    /// order](https://webassembly.github.io/spec/core/binary/modules.html#binary-module),
-    /// and to ensure that each kind of section (other than custom sections) is
-    /// only defined once. While this is a potential footgun, it also allows you
-    /// to use this crate to easily construct test cases for bad Wasm module
-    /// encodings.
-    pub fn section(&mut self, section: &impl Section) -> &mut Self {
-        self.bytes.push(section.id());
-        section.encode(&mut self.bytes);
-        self
-    }
-
-    /// Get the encoded Wasm module as a slice.
-    pub fn as_slice(&self) -> &[u8] {
-        &self.bytes
-    }
-
-    /// Finish writing this Wasm module and extract ownership of the encoded
-    /// bytes.
-    pub fn finish(self) -> Vec<u8> {
-        self.bytes
-    }
-}
-
-impl Default for Module {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Represents a WebAssembly component that is being encoded.
-///
-/// Unlike core WebAssembly modules, the sections of a component
-/// may appear in any order and may be repeated.
-#[derive(Clone, Debug)]
-pub struct Component {
-    bytes: Vec<u8>,
-}
-
-impl Component {
-    /// Begin writing a new `Component`.
-    pub fn new() -> Self {
-        Self {
-            bytes: vec![
-                0x00, 0x61, 0x73, 0x6D, // magic (`\0asm`)
-                0x0a, 0x00, 0x01, 0x00, // version
-            ],
-        }
-    }
-
-    /// Finish writing this component and extract ownership of the encoded bytes.
-    pub fn finish(self) -> Vec<u8> {
-        self.bytes
-    }
-
-    /// Write a section to this component.
-    pub fn section(&mut self, section: &impl ComponentSection) -> &mut Self {
-        self.bytes.push(section.id());
-        section.encode(&mut self.bytes);
-        self
-    }
-}
-
-impl Default for Component {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Known section identifiers of WebAssembly modules.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd)]
-#[repr(u8)]
-pub enum SectionId {
-    /// The custom section.
-    Custom = 0,
-    /// The type section.
-    Type = 1,
-    /// The import section.
-    Import = 2,
-    /// The function section.
-    Function = 3,
-    /// The table section.
-    Table = 4,
-    /// The memory section.
-    Memory = 5,
-    /// The global section.
-    Global = 6,
-    /// The export section.
-    Export = 7,
-    /// The start section.
-    Start = 8,
-    /// The element section.
-    Element = 9,
-    /// The code section.
-    Code = 10,
-    /// The data section.
-    Data = 11,
-    /// The data count section.
-    DataCount = 12,
-    /// The tag section.
-    ///
-    /// This section is supported by the exception handling proposal.
-    Tag = 13,
-}
-
-impl From<SectionId> for u8 {
-    #[inline]
-    fn from(id: SectionId) -> u8 {
-        id as u8
-    }
-}
-
-/// Known section identifiers of WebAssembly components.
-///
-/// These sections are supported by the component model proposal.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd)]
-#[repr(u8)]
-pub enum ComponentSectionId {
-    /// The section is a custom section.
-    Custom = 0,
-    /// The section is a type section.
-    Type = 1,
-    /// The section is an import section.
-    Import = 2,
-    /// The section is a function section.
-    Function = 3,
-    /// The section is a module section.
-    Module = 4,
-    /// The section is a component section.
-    Component = 5,
-    /// The section is an instance section.
-    Instance = 6,
-    /// The section is an export section.
-    Export = 7,
-    /// The section is a start section.
-    Start = 8,
-    /// The section is an alias section.
-    Alias = 9,
-}
-
-impl From<ComponentSectionId> for u8 {
-    #[inline]
-    fn from(id: ComponentSectionId) -> u8 {
-        id as u8
-    }
-}
-
-/// The type of an entity.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum EntityType {
-    /// A function type.
-    ///
-    /// The value is an index into the types section.
-    Function(u32),
-    /// A table type.
-    Table(TableType),
-    /// A memory type.
-    Memory(MemoryType),
-    /// A global type.
-    Global(GlobalType),
-    /// A tag type.
-    ///
-    /// This variant is used with the exception handling proposal.
-    Tag(TagType),
-}
-
-impl EntityType {
-    fn encode(&self, bytes: &mut Vec<u8>) {
-        match self {
-            Self::Function(i) => {
-                bytes.push(0x00);
-                bytes.extend(encoders::u32(*i));
-            }
-            Self::Table(t) => {
-                bytes.push(0x01);
-                t.encode(bytes);
-            }
-            Self::Memory(t) => {
-                bytes.push(0x02);
-                t.encode(bytes);
-            }
-            Self::Global(t) => {
-                bytes.push(0x03);
-                t.encode(bytes);
-            }
-            Self::Tag(t) => {
-                bytes.push(0x04);
-                t.encode(bytes);
-            }
-        }
-    }
-}
-
-impl From<TableType> for EntityType {
-    fn from(t: TableType) -> Self {
-        Self::Table(t)
-    }
-}
-
-impl From<MemoryType> for EntityType {
-    fn from(t: MemoryType) -> Self {
-        Self::Memory(t)
-    }
-}
-
-impl From<GlobalType> for EntityType {
-    fn from(t: GlobalType) -> Self {
-        Self::Global(t)
-    }
-}
-
-impl From<TagType> for EntityType {
-    fn from(t: TagType) -> Self {
-        Self::Tag(t)
-    }
-}
 
 #[cfg(test)]
 mod test {
